requirements.txt
python-dotenv==0.15.0
slack-bolt==1.0.0
flask==2.0.1
werkzeug==2.3.7